some more steps toward MVC refactor, not working right now
